requirements.txt
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
certifi==2023.11.17
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
colorama==0.4.6
comtypes==1.2.0
decorator==4.4.2
distro==1.9.0
exceptiongroup==1.2.0
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.13.1
fsspec==2023.12.2
future==0.18.3
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==1.0.2
httpx==0.26.0
hyperframe==5.2.0
idna==2.10
imageio==2.33.1
imageio-ffmpeg==0.4.9
Jinja2==3.1.2
llvmlite==0.41.1
MarkupSafe==2.1.3
more-itertools==10.1.0
moviepy==1.0.3
mpmath==1.3.0
mutagen==1.47.0
networkx==3.2.1
numba==0.58.1
numpy==1.26.2
openai==1.7.2
openai-whisper==20231117
opencv-python==4.8.1.78
outcome==1.3.0.post0
packaging==23.2
Pillow==10.1.0
praw==7.7.1
prawcore==2.4.0
proglog==0.1.10
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pypiwin32==223
PySocks==1.7.1
python-dotenv==1.0.0
pyttsx3 @ git+https://github.com/nateshmbhat/pyttsx3@5d3755b060a980f48fcaf81df018dd06cbd17a8f
pytube==15.0.0
pywin32==306
regex==2023.10.3
requests==2.31.0
rfc3986==1.5.0
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
subprocess.run==0.0.8
sympy==1.12
tiktoken==0.5.2
torch==2.1.2
tqdm==4.66.1
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
update-checker==0.18.0
urllib3==2.1.0
websocket-client==1.7.0
wsproto==1.2.0
